Commit files with `--editor`
This launches `git commit` with the `--editor` option which allows a
user to review changes before the entire operation goes through. It also
allows them to cancel the process by forcing their editor to exit with a
non-zero exit code (for example `:cq` in Vim).

This is useful in case some set of unexpected changes get pulled in from
pay-server and the operation needs to be aborted.